requirements.txt
mysql-connector-python==8.3.0
pillow==10.2.0
